minimega: shift code
From misc to netconfig since it was only used in netconfig.
